refactor: Service Interface로 추상화 (#57)
