+++ src/main/java/com/meommu/meommuapi/core/auth/repository/RefreshTokenRepository.java
@@ -1,7 +1,5 @@
 package com.meommu.meommuapi.core.auth.repository;
 
-import org.springframework.stereotype.Repository;
-
 public interface RefreshTokenRepository {
 
 	void save(Long userId, String refreshToken, Long milliseconds);
