+++ src/main/java/com/meommu/meommuapi/core/diary/dto/DiarySearchCriteria.java
@@ -1,7 +1,5 @@
 package com.meommu.meommuapi.core.diary.dto;
 
-import java.time.LocalDate;
-
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Getter;
