+++ src/main/java/com/meommu/meommuapi/core/image/exception/ImageNotFoundException.java
@@ -1,7 +1,7 @@
 package com.meommu.meommuapi.core.image.exception;
 
-import com.meommu.meommuapi.global.exception.NotFoundException;
 import com.meommu.meommuapi.core.image.exception.errorCode.ImageErrorCode;
+import com.meommu.meommuapi.global.exception.NotFoundException;
 
 public class ImageNotFoundException extends NotFoundException {
