+++ src/main/java/com/meommu/meommuapi/core/image/exception/S3DeleteException.java
@@ -1,7 +1,7 @@
 package com.meommu.meommuapi.core.image.exception;
 
-import com.meommu.meommuapi.global.exception.InternalServerException;
 import com.meommu.meommuapi.core.image.exception.errorCode.ImageErrorCode;
+import com.meommu.meommuapi.global.exception.InternalServerException;
 
 public class S3DeleteException extends InternalServerException {
 	public S3DeleteException() {
