+++ src/main/java/com/meommu/meommuapi/core/image/exception/S3UploadException.java
@@ -1,7 +1,7 @@
 package com.meommu.meommuapi.core.image.exception;
 
-import com.meommu.meommuapi.global.exception.InternalServerException;
 import com.meommu.meommuapi.core.image.exception.errorCode.ImageErrorCode;
+import com.meommu.meommuapi.global.exception.InternalServerException;
 
 public class S3UploadException extends InternalServerException {
 	public S3UploadException() {
