+++ src/main/java/com/meommu/meommuapi/core/kindergarten/controller/KindergartenController.java
@@ -13,11 +13,11 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.meommu.meommuapi.core.auth.dto.AuthInfo;
 import com.meommu.meommuapi.core.auth.token.Auth;
-import com.meommu.meommuapi.core.kindergarten.dto.KindergartenPasswordUpdateRequest;
 import com.meommu.meommuapi.core.kindergarten.dto.EmailRequest;
-import com.meommu.meommuapi.core.kindergarten.dto.MyInfoResponse;
-import com.meommu.meommuapi.core.kindergarten.dto.KindergartenUpdateRequest;
+import com.meommu.meommuapi.core.kindergarten.dto.KindergartenPasswordUpdateRequest;
 import com.meommu.meommuapi.core.kindergarten.dto.KindergartenResponse;
+import com.meommu.meommuapi.core.kindergarten.dto.KindergartenUpdateRequest;
+import com.meommu.meommuapi.core.kindergarten.dto.MyInfoResponse;
 import com.meommu.meommuapi.core.kindergarten.dto.SignUpRequest;
 import com.meommu.meommuapi.core.kindergarten.service.KindergartenService;
