+++ src/main/java/com/meommu/meommuapi/core/kindergarten/exception/EmailCodeExpiredException.java
@@ -1,7 +1,7 @@
 package com.meommu.meommuapi.core.kindergarten.exception;
 
-import com.meommu.meommuapi.global.exception.BadRequestException;
 import com.meommu.meommuapi.core.kindergarten.exception.errorCode.KindergartenErrorCode;
+import com.meommu.meommuapi.global.exception.BadRequestException;
 
 public class EmailCodeExpiredException extends BadRequestException {
