+++ src/main/java/com/meommu/meommuapi/core/kindergarten/repository/EmailCodeRepository.java
@@ -1,7 +1,5 @@
 package com.meommu.meommuapi.core.kindergarten.repository;
 
-import org.springframework.stereotype.Repository;
-
 public interface EmailCodeRepository {
 
 	void save(String email, String code, Long milliseconds);
