+++ src/main/java/com/meommu/meommuapi/core/proxy/service/ProxyServiceImpl.java
@@ -1,0 +1,26 @@
+package com.meommu.meommuapi.core.proxy.service;
+
+import java.net.HttpURLConnection;
+import java.net.URL;
+
+import org.springframework.stereotype.Service;
+
+import com.amazonaws.util.IOUtils;
+import com.meommu.meommuapi.core.proxy.dto.Html2canvasRequest;
+import com.meommu.meommuapi.core.proxy.exception.Html2canvasInternalException;
+
+@Service
+public class ProxyServiceImpl implements ProxyService {
+
+	@Override
+	public byte[] html2canvasProxy(Html2canvasRequest request) {
+		try {
+			URL targetUrl = new URL(request.getUrl());
+			HttpURLConnection connection = (HttpURLConnection)targetUrl.openConnection();
+			connection.setRequestMethod("GET");
+			return IOUtils.toByteArray(connection.getInputStream());
+		} catch (Exception e) {
+			throw new Html2canvasInternalException();
+		}
+	}
+}
